Move RoomDatabaseExtensions to data
