+++ wallet/src/de/schildbach/wallet/data/DashPayContactRequestDao.kt
@@ -5,7 +5,6 @@ import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import de.schildbach.wallet.getDistinct
 
 @Dao
 interface DashPayContactRequestDao {
